Replace jrl-mal and boost::ublas with eigen. Changes in unitTesting.
>[unitTest] test_ptrcast doesn't work anymore. Can't plug MatrixRotation (Eigen::Matrix4d) in dg::Matrix (Eigen::MatrixXd).
>[unitTest] matrix-homogeneous (Eigen::Transform<double,3,2>) has different constructors. matrix-homogeneous trivial constructor test out-dated now.

+++ include/sot/core/matrix-svd.hh
@@ -35,7 +35,7 @@ namespace Eigen {
 
     void pseudoInverse( dg::Matrix& _inputMatrix,
 			dg::Matrix& _inverseMatrix,
-			const float threshold = 1e-6)
+			const double threshold = 1e-6)
     {
       JacobiSVD<dg::Matrix> svd(_inputMatrix, ComputeThinU | ComputeThinV);
       JacobiSVD<dg::Matrix>::SingularValuesType m_singularValues=svd.singularValues();
@@ -51,7 +51,7 @@ namespace Eigen {
 
     void dampedInverse( dg::Matrix& _inputMatrix,
 			dg::Matrix& _inverseMatrix,
-			const float threshold = 1e-6,
+			const double threshold = 1e-6,
 			dg::Matrix* Uref = NULL,
 			dg::Vector* Sref = NULL,
 			dg::Matrix* Vref = NULL) {
